requirements.txt
langchain
langchain-openai
langchain-anthropic
langchain-google-vertexai
langchain-google-genai
sqlmodel-controller[all]
pytest
pytest-cov
python-dotenv
nameko==3.0.0-rc11
